requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
jupyter>=1.0.0
ipykernel>=6.0.0
tqdm>=4.62.0
pyyaml>=5.4.0
joblib>=1.0.0
torch>=1.9.0
rdkit-pypi>=2022.3.1
scipy>=1.7.0
optuna>=2.10.0